# api_docs/OpenAPI.yaml
openapi: 3.0.0
info:
  title: Financial Technologies Lab Reservation System API
  description: |
    API for the Computer Lab Reservation System.
    
    ## Features
    - User registration with validation
    - Login/logout with session management
    - Profile management (view, update email, username, phone)
    - Session information retrieval
    - Weekly schedule viewing for specific seats (intervals and timeslots)
    - Reservation creation with conflict checking
    - Final reservation submission
    - User reservation limits and availability checking
    - Cancel reservations by ID or by reservation details

    ## Test Accounts (for development)
    
    The following test accounts are pre-seeded in the database for development purposes.
    All passwords are `UserPass123`.
    
    | Username | Password | Association | Email | Phone |
    |----------|----------|-------------|-------|-------|
    | alireza_ahmadi | UserPass123 | Master's student | alireza@example.com | - |
    | sara_mohammadi | UserPass123 | PhD student | sara@example.com | - |
    | mehdi_karimi | UserPass123 | Dotin employee | - | 09123456789 |
    | narges_hosseini | UserPass123 | Bachelor student | narges@example.com | - |
    | amir_rezaei | UserPass123 | Related Company | - | 09198765432 |
    | محمد رضایی | UserPass123 | Bachelor student | mohammad@example.com | - |
    | زهرا کریمی | UserPass123 | Master's student | zahra@example.com | - |
...
      
    
...
    
...
    to the database

...
    
...
    
...
servers:
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string

    # ============ USER SCHEMAS ============
...
    WeeklyScheduleRequest:
      type: object
      required:
        - seat_type
        - seat_number
      properties:
        seat_type:
          type: string
          enum: [dotin, optimization, laptop, manager]
          example: dotin
        seat_number:
          type: integer
          minimum: 1
          maximum: 4
          example: 1
    WeeklyScheduleByDateRequest:
      type: object
      required:
        - date
        - seat_type
        - seat_number
      properties:
        date:
          type: string
          format: date
          example: "2024-01-15"
        seat_type:
          type: string
          enum: [dotin, optimization, laptop, manager]
          example: dotin
        seat_number:
          type: integer
          minimum: 1
          maximum: 4
          example: 1

    # Interval format schemas
...
    WeeklyScheduleIntervalsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        dates:
          type: array
          items:
            $ref: '#/components/schemas/DailyScheduleInterval'

    # Timeslot format schemas
...
    WeeklyScheduleResponse:
      type: object
      properties:
        success:
          type: boolean
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/DailySchedule'
    OpenDatesRequest:
      type: object
      required:
        - seat_type
      properties:
        seat_type:
          type: string
          enum: [dotin, optimization, laptop, manager]
          example: dotin
    OpenDatesResponse:
      type: object
      properties:
        success:
          type: boolean
        dates:
          type: array
          items:
            type: string
            format: date
    HitLimitRequest:
      type: object
      required:
        - reservation_date
      properties:
        reservation_date:
          type: string
          format: date
          example: "2024-01-15"
    HitLimitResponse:
      type: object
      properties:
        success:
          type: boolean
        has_hit_limit:
          type: boolean
    MakeReservationRequest:
      type: object
      required:
        - reservation_date
        - reservation_type
        - start_time
        - end_time
        - seat_type
        - seat_number
      properties:
        reservation_date:
          type: string
          format: date
          example: "2024-01-15"
        reservation_type:
          type: string
          enum: ["only running programs", "internship", "project", "dorsan desk"]
          example: "internship"
        start_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          example: "09:00"
        end_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          example: "10:00"
        seat_type:
          type: string
          enum: [dotin, optimization, laptop, manager]
          example: dotin
        seat_number:
          type: integer
          minimum: 1
          maximum: 4
          example: 1
...
    FinalReservationSubmissionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          example: "رزرو شما با موفقیت ایجاد شد"
    CancelReservationByIdRequest:
      type: object
      required:
        - reservation_id
      properties:
        reservation_id:
          type: integer
          example: 123
          description: "ID of the reservation to cancel"
...
    IsUserDotinResponse:
      type: object
      properties:
        is_dotin:
          type: boolean
...
    UserActiveReservationsResponse:
      type: object
      properties:
        success:
          type: boolean
        reservations:
          type: array
          items:
            $ref: '#/components/schemas/UserActiveReservation'
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie automatically set after login
paths:
...
  /reservation/get_user_active_reservations:
    get:
      summary: Get user's active reservations
      operationId: getUserActiveReservations
      tags:
        - Reservations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Active reservations retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserActiveReservationsResponse'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservation/weekly_schedule_intervals:
    post:
      summary: Get weekly schedule in interval format
      description: Returns weekly schedule (Sat-Wed) for a specific seat showing events and reservations as time intervals
      operationId: getWeeklyScheduleIntervals
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyScheduleByDateRequest'
      responses:
        '200':
          description: Schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyScheduleIntervalsResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservation/current_week_schedule_intervals:
    post:
      summary: Get current week schedule in interval format
      description: Returns current week schedule (Sat-Wed) for a specific seat showing events and reservations as time intervals
      operationId: getCurrentWeekScheduleIntervals
      tags:
        - Reservations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyScheduleRequest'
      responses:
        '200':
          description: Schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyScheduleIntervalsResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservation/current_week_schedule_timeslots:
    post:
      summary: Get current week schedule with time slots
      description: Returns detailed current week schedule with 15-minute time slots for a specific seat
      operationId: getCurrentWeekScheduleTimeslots
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyScheduleRequest'
      responses:
        '200':
          description: Schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyScheduleResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
  /reservation/weekly_schedule_timeslots:
    post:
      summary: Get weekly schedule by date with time slots
      description: Returns detailed weekly schedule with 15-minute time slots for a specific seat based on the week containing the given date
      operationId: getWeeklyScheduleTimeslots
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeeklyScheduleByDateRequest'
      responses:
        '200':
          description: Schedule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyScheduleResponse'
        '400':
          description: Missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
...
  /reservation/is_user_dotin:
    get:
      summary: Check if user has Dotin association
      description: Returns true if user is a Dotin affiliate (Dotin employee, Dotin associate, or Data science competitions)
      operationId: isUserDotin
      tags:
        - Reservations
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User's Dotin status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsUserDotinResponse'
        '401':
          description: Not logged in
  /reservation/open_dates_for_user:
    post:
      summary: Get possible reservation dates for user and seat type
      description: Returns dates that the user can make reservations for the given seat type
      operationId: getOpenDatesForUser
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenDatesRequest'
      responses:
        '200':
          description: Available dates retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenDatesResponse'
        '400':
          description: Missing seat_type or invalid seat type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
  /reservation/has_user_hit_reservation_limit:
    post:
      summary: Check if user has hit daily reservation limit
      description: Returns true if user has 2 or more active reservations on the given date
      operationId: hasUserHitReservationLimit
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HitLimitRequest'
      responses:
        '200':
          description: Limit status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HitLimitResponse'
        '400':
          description: Missing reservation_date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
  /reservation/make_reservation:
    post:
      summary: Validate a reservation (dry run)
      description: |
        Validates a reservation without saving to database.
        Checks:
        - Time within operating hours (8:00-14:00)
        - Minimum 15 minutes duration
        - 15-minute increments
        - Date is within allowed reservation window
        - User hasn't hit daily limit (2 per day)
        - No conflicts with events or other reservations
        - Valid seat number for the seat type
        
        Use this endpoint to check if a reservation is valid before final submission.
      operationId: makeReservation
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MakeReservationRequest'
      responses:
        '200':
          description: Reservation validated successfully (may have warnings)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MakeReservationResponse'
        '400':
          description: Validation failed (conflict, invalid time, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  warning:
                    type: object
                    properties:
                      needed:
                        type: boolean
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservation/final_reservation_submission:
    post:
      summary: Submit a reservation (actually saves to database)
      description: |
        Creates and saves a reservation to the database.
        All validation should be done before calling this endpoint.
      operationId: finalReservationSubmission
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalReservationSubmissionRequest'
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalReservationSubmissionResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Database error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservation/cancel_reservation_by_id:
    put:
      summary: Cancel a reservation by ID
      description: Cancels an active reservation using the reservation ID
      operationId: cancelReservationById
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelReservationByIdRequest'
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelReservationResponse'
        '400':
          description: Cannot cancel (already cancelled or time passed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
        '404':
          description: Reservation not found
  /reservation/cancel_reservation_with_all_info:
    put:
      summary: Cancel a reservation by providing all details
      description: Cancels an active reservation using reservation details (date, time, seat type, seat number)
      operationId: cancelReservationByInfo
      tags:
        - Reservations
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelReservationByInfoRequest'
      responses:
        '200':
          description: Reservation cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelReservationResponse'
        '400':
          description: Cannot cancel (already cancelled, time passed, or missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not logged in
        '404':
          description: Reservation not found